Split LSP lint scheduling and diagnostics push into server_diagnostics.go
Move the debounce machinery (scheduleLint, pendingLint, runLintIfCurrent,
stopPendingLints), the run-mode gate, clearOpenDiagnostics, runLint, and
its diagnostic-routing helpers (resolveMaxInputBytes,
surfaceForeignDiagnostics, messageTypeForLint, partitionDocDiagnostics)
out of server.go. The LSP wire-shape conversion stays in diagnostics.go.
Pure file-split refactor — no behavior change. server.go: 1756 -> 1362.
